chore: extract duplicated code into separate files
