+++ src/colorMapper.js
@@ -1,3 +1,5 @@
+import { generateColorVector } from './colorUtils'
+
 function pickHex (color1, color2, weight) {
     const w1 = weight
     const w2 = 1 - w1
@@ -5,30 +7,6 @@ function pickHex (color1, color2, weight) {
         Math.round(color1[1] * w1 + color2[1] * w2),
         Math.round(color1[2] * w1 + color2[2] * w2)]
     return rgb
-}
-
-function generateHash (name) {
-    // Return a vector (0.0->1.0) that is a hash of the input string.
-    // The hash is computed to favor early characters over later ones, so
-    // that strings with similar starts have similar vectors. Only the first
-    // 6 characters are considered.
-    const MAX_CHAR = 6
-
-    var hash = 0
-    var maxHash = 0
-    var weight = 1
-    var mod = 10
-
-    if (name) {
-        for (var i = 0; i < name.length; i++) {
-            if (i > MAX_CHAR) { break }
-            hash += weight * (name.charCodeAt(i) % mod)
-            maxHash += weight * (mod - 1)
-            weight *= 0.70
-        }
-        if (maxHash > 0) { hash = hash / maxHash }
-    }
-    return hash
 }
 
 export function allocationColorMapper (d, originalColor) {
@@ -44,15 +22,8 @@ export function allocationColorMapper (d, originalColor) {
 export function offCpuColorMapper (d, originalColor) {
     if (d.highlight) return originalColor
 
-    let name = d.data.n || d.data.name
-    let vector = 0
-    const nameArr = name.split('`')
-
-    if (nameArr.length > 1) {
-        name = nameArr[nameArr.length - 1] // drop module name if present
-    }
-    name = name.split('(')[0] // drop extra info
-    vector = generateHash(name)
+    const name = d.data.n || d.data.name
+    const vector = generateColorVector(name)
 
     const r = 0 + Math.round(55 * (1 - vector))
     const g = 0 + Math.round(230 * (1 - vector))
